replace stat with openssl md5
Command line arguments for stat are different on macOS (Darwin) and
Linux. If the file does not end with a empty line some editors (like
nano) will add one even if you save the file without any modifications.
In this case, the md5 hash is different even if the user did not change
any content. Adding a newline is a workaround for this.